third_party/WebKit/LayoutTests/media/controls/download-button-displays-with-preload-none.html

<!DOCTYPE html>
<title>media controls download button preload none</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../media-file.js"></script>
<script src="../media-controls.js"></script>
<video controls preload="none" src="https://someexample.example/example.mp4"></video>
<script>
async_test(function(t) {
  var video = document.querySelector("video");

  testRunner.layoutAndPaintAsyncThen(t.step_func_done(function() {
    assert_true(isVisible(downloadButton(video)));
  }));

  function downloadButton(videoElement) {
    var controlID = '-internal-media-controls-download-button';
    var button = mediaControlsElement(
        window.internals.shadowRoot(videoElement).firstChild, controlID);
    if (!button)
      throw 'Failed to find download button';
    return button;
  }
});
</script>
